Refactor: "이름 변경"(InputManager -> ConsoleInputManager)
